Further renaming work, specifically renaming the file and updating other references.
